无敌菜单

+++ Trace/AIO.h
@@ -12,11 +12,11 @@ public:
 	}
 
 
-	void GankTips(tuple<bool, float, int, int>& v);
+	void GankTips(tuple<bool, int, int, int>& v);
 
 	void DrawCD();
 
-	void DrawPath(tuple<bool, float, int>& v);
+	void DrawPath(tuple<bool, int, int>& v);
 
 
 	static void Present();
